Separate func by type
