+++ pkg/key/appcatalog.go
@@ -7,7 +7,31 @@ import (
 	"github.com/giantswarm/app/v2/pkg/label"
 )
 
-func CatalogType(customResource v1alpha1.AppCatalog) string {
+func AppCatalogTitle(customResource v1alpha1.AppCatalog) string {
+	return customResource.Spec.Title
+}
+
+func AppCatalogStorageURL(customResource v1alpha1.AppCatalog) string {
+	return customResource.Spec.Storage.URL
+}
+
+func AppCatalogConfigMapName(customResource v1alpha1.AppCatalog) string {
+	return customResource.Spec.Config.ConfigMap.Name
+}
+
+func AppCatalogConfigMapNamespace(customResource v1alpha1.AppCatalog) string {
+	return customResource.Spec.Config.ConfigMap.Namespace
+}
+
+func AppCatalogSecretName(customResource v1alpha1.AppCatalog) string {
+	return customResource.Spec.Config.Secret.Name
+}
+
+func AppCatalogSecretNamespace(customResource v1alpha1.AppCatalog) string {
+	return customResource.Spec.Config.Secret.Namespace
+}
+
+func AppCatalogType(customResource v1alpha1.AppCatalog) string {
 	if val, ok := customResource.ObjectMeta.Labels[label.CatalogType]; ok {
 		return val
 	}
@@ -15,7 +39,7 @@ func CatalogType(customResource v1alpha1.AppCatalog) string {
 	return ""
 }
 
-func CatalogVisibility(customResource v1alpha1.AppCatalog) string {
+func AppCatalogVisibility(customResource v1alpha1.AppCatalog) string {
 	if val, ok := customResource.ObjectMeta.Labels[label.CatalogVisibility]; ok {
 		return val
 	}
